Allow editing of events and adding new events on overview page

+++ js/upcoming/upcoming.controller.js
@@ -31,6 +31,10 @@
      * @param evt
      */
     $scope.removeEvent = function(evt) {
+      if(evt.name === ''){
+        // Blank event from adding a new row
+        return loadEvents();
+      }
       if(window.confirm('Are you sure you want to remove this event?')) {
         $scope.$eval(function() {
           VSeats.eventService.remove(evt,
@@ -45,6 +49,37 @@
           );
         });
       }
+    };
+
+    /**
+     * Update Event
+     * @param evt
+     * @param newData
+     */
+    $scope.updateEvent = function(evt, newData) {
+      // Update the event data with the new data values
+      evt.name = newData.name;
+      evt.date = newData.date;
+      evt.venue.name = newData.venuename;
+      evt.venue.city = newData.city;
+      evt.venue.state = newData.state;
+
+      if(evt.name === '') {
+        // Blank event data trying to be saved
+        return;
+      }
+
+      // Update the event with the API
+      VSeats.eventService.update(evt,
+        function(msg) {
+          toaster.pop('success', "Success", "The event was updated successfully!");
+          loadEvents();
+        },
+        function(msg) {
+          toaster.pop('error', "Error", "There was an error updating the event!");
+          loadEvents();
+        }
+      );
     };
 
     /**
